# src/client/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DUCC Website API",
    "description": "API documentation for the Durham University Canoe Club website.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session cookie used for authentication."
      }
    }
  },
  "paths": {
...
    "/api/event/{id}/isOnWaitlist": {
        "get": {
            "summary": "Check waitlist status",
            "description": "Returns whether the current user is on the waitlist for the event.",
            "tags": ["Events"],
            "security": [{ "cookieAuth": [] }],
            "parameters": [{ "name": "id", "in": "path", "required": true, "description": "The ID of the event.", "schema": { "type": "integer" } }],
            "responses": { "200": { "description": "Status returned" } }
        }
    },
    "/api/event/{id}/waitlist/join": {
        "post": {
            "summary": "Join waitlist",
            "description": "Adds the current user to the waitlist for a full event.",
            "tags": ["Events"],
            "security": [{ "cookieAuth": [] }],
            "parameters": [{ "name": "id", "in": "path", "required": true, "description": "The ID of the event.", "schema": { "type": "integer" } }],
            "responses": { "200": { "description": "Joined successfully" } }
        }
    },
    "/api/event/{id}/waitlist/leave": {
        "post": {
            "summary": "Leave waitlist",
            "description": "Removes the current user from the event's waitlist.",
            "tags": ["Events"],
            "security": [{ "cookieAuth": [] }],
            "parameters": [{ "name": "id", "in": "path", "required": true, "description": "The ID of the event.", "schema": { "type": "integer" } }],
            "responses": { "200": { "description": "Left successfully" } }
        }
    },
    "/api/event/{id}/waitlist": {
        "get": {
            "summary": "Get waitlist info",
            "description": "Retrieves the current waitlist for an event.",
            "tags": ["Events"],
            "parameters": [{ "name": "id", "in": "path", "required": true, "description": "The ID of the event.", "schema": { "type": "integer" } }],
            "responses": { "200": { "description": "Waitlist data" } }
        }
    },
...
  }
}